+++ lib/logger/backends/json_dummy_encoder.ex
@@ -1,0 +1,3 @@
+defmodule Logger.Backends.JSON.DummyEncoder do
+  def encode(object), do: {:ok, inspect(object)}
+end