+++ CHANGELOG.md
@@ -6,3 +6,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+- base json backend implementation